Refactor, updates -Added date selection for Log JSON generation (tab 2) -Added fixed column amount tables support -Added feature to the mover to avoid duplicates

+++ reader/Reader.java
@@ -2,53 +2,59 @@ package reader;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
 
 public class Reader {
 
-	private String prefix="";
-	private String separator="_";
+	private String prefix = "";
+	private String separator = "_";
+	private String DUPLICATE_FILE_SUFFIX = " (#)";
 	private String sourceFolder;
 	private String destFolder;
-	
-	public Reader(String sourceFolder, String destFolder){
+
+	public Reader(String sourceFolder, String destFolder) {
 		this.sourceFolder = sourceFolder;
 		this.destFolder = destFolder;
 	}
-	public Reader(String sourceFolder, String destFolder, String prefix){
+
+	public Reader(String sourceFolder, String destFolder, String prefix) {
 		this.sourceFolder = sourceFolder;
 		this.destFolder = destFolder;
 		this.prefix = prefix;
 	}
-	public Reader(String sourceFolder, String destFolder, String prefix, String separator){
+
+	public Reader(String sourceFolder, String destFolder, String prefix, String separator) {
 		this.sourceFolder = sourceFolder;
 		this.destFolder = destFolder;
 		this.prefix = prefix;
 		this.separator = separator;
 	}
-	
-	public String parseFolder() throws IOException{
+
+	public String parseFolder() throws IOException {
 		StringBuilder result = new StringBuilder();
 		File folder = new File(sourceFolder);
-		if (folder.isDirectory()){
+		if (folder.isDirectory()) {
 			String[] list = folder.list();
-			for (int i=0; i<list.length; ++i){
+			for (int i = 0; i < list.length; ++i) {
 				String name = list[i];
-				if (name.startsWith(prefix)){
-					
+				if (name.startsWith(prefix) && !isNameDuplicate(name)) {
 					StringBuilder destPathBuilder = new StringBuilder();
 					String nameWithoutPrefix = name.substring(prefix.length());
-					
+
 					String[] path = nameWithoutPrefix.split(separator);
-					for (int j=0; j<path.length; ++j){
+					for (int j = 0; j < path.length; ++j) {
 						destPathBuilder.append(File.separator);
 						destPathBuilder.append(path[j]);
 					}
 					String newPath = destFolder + destPathBuilder.toString();
-					
+
 					File file = new File(sourceFolder + File.separator + name);
+					removeDuplicates(file, list);
 					File newFile = new File(newPath);
 					newFile.getParentFile().mkdirs();
-					if (newFile.exists()){
+					if (newFile.exists()) {
 						newFile.delete();
 					}
 					cleanUp(newFile.getName(), destFolder);
@@ -56,24 +62,72 @@ public class Reader {
 					file.renameTo(newFile);
 				}
 			}
-			
+
 		}
 		return result.toString();
 	}
-	
-	public void cleanUp(String fileName, String location){
+
+	private boolean isNameDuplicate(String name) {
+		String firstPart = " (";
+		String secondPart = ").txt";
+		int firstPartIndex = name.lastIndexOf(firstPart);
+		int secondPartIndex = name.lastIndexOf(secondPart);
+		if (name.contains(firstPart) && name.endsWith(secondPart)
+				&& isNumber(name.substring(firstPartIndex + 2, secondPartIndex))) {
+			return true;
+		}
+		return false;
+	}
+
+	private void removeDuplicates(File file, String[] fileList) {
+		String fileNameWithoutExtension = file.getName().replace(".txt", "");
+		List<File> duplicates = new ArrayList<File>();
+		for (int i = 0; i < fileList.length; ++i) {
+			if (fileList[i].startsWith(fileNameWithoutExtension)) {
+				duplicates.add(new File(file.getParentFile() + File.separator + fileList[i]));
+			}
+		}
+		if (duplicates.size() > 1) {
+			Iterator<File> it = duplicates.iterator();
+			String goodFileName = fileNameWithoutExtension
+					+ DUPLICATE_FILE_SUFFIX.replace("#", duplicates.size() - 1 + "") + ".txt";
+			File goodFile = null;
+			while (it.hasNext()) {
+				File currFile = it.next();
+				if (!currFile.getName().equals(goodFileName)) {
+					currFile.delete();
+				} else {
+					goodFile = currFile;
+				}
+			}
+			if (goodFile != null) {
+				goodFile.renameTo(new File(file.getParentFile() + File.separator + fileNameWithoutExtension + ".txt"));
+			}
+		}
+	}
+
+	public void cleanUp(String fileName, String location) {
 		File rootFolder = new File(location);
-		if (rootFolder.isDirectory()){
+		if (rootFolder.isDirectory()) {
 			File[] contents = rootFolder.listFiles();
-			for (int i=0; i<contents.length; ++i){
-				if (contents[i].isDirectory()){
+			for (int i = 0; i < contents.length; ++i) {
+				if (contents[i].isDirectory()) {
 					cleanUp(fileName, contents[i].getPath());
-				}else {
-					if (contents[i].getName().equals(fileName)){
+				} else {
+					if (contents[i].getName().equals(fileName)) {
 						contents[i].delete();
 					}
 				}
 			}
+		}
+	}
+
+	private boolean isNumber(String value) {
+		try {
+			Integer.parseInt(value);
+			return true;
+		} catch (NumberFormatException e) {
+			return false;
 		}
 	}
 }
